Deprecate org.http4s.util.threads
