+++ blaze-client/src/test/scala/org/http4s/client/blaze/BlazeHttp1ClientSpec.scala
@@ -9,7 +9,7 @@ package client
 package blaze
 
 import cats.effect.IO
-import org.http4s.util.threads.newDaemonPoolExecutionContext
+import org.http4s.internal.threads.newDaemonPoolExecutionContext
 
 class BlazeHttp1ClientSpec extends ClientRouteTestBattery("BlazeClient") {
   def clientResource =
