+++ core/src/main/scala/org/http4s/util/threads.scala
@@ -6,12 +6,10 @@
 
 package org.http4s.util
 
-import java.lang.Thread.UncaughtExceptionHandler
 import java.util.concurrent._
-import java.util.concurrent.atomic.AtomicLong
 import scala.concurrent.ExecutionContext
-import scala.concurrent.duration._
 
+@deprecated("Not related to HTTP. Will be removed from public API.", "0.21.5")
 object threads {
   final case class ThreadPriority(toInt: Int)
   case object ThreadPriority {
@@ -28,27 +26,15 @@ object threads {
       priority: ThreadPriority = ThreadPriority.Norm,
       uncaughtExceptionHandler: PartialFunction[(Thread, Throwable), Unit] = PartialFunction.empty,
       backingThreadFactory: ThreadFactory = Executors.defaultThreadFactory
-  ): ThreadFactory =
-    new ThreadFactory {
-      val count = new AtomicLong(0)
-      override def newThread(r: Runnable): Thread = {
-        val thread = backingThreadFactory.newThread(r)
-        thread.setName(name(count.getAndIncrement))
-        thread.setDaemon(daemon)
-        thread.setPriority(priority.toInt)
-        thread.setUncaughtExceptionHandler(new UncaughtExceptionHandler {
-          override def uncaughtException(t: Thread, e: Throwable): Unit =
-            uncaughtExceptionHandler.orElse(fallthrough)((t, e))
-          val fallthrough: PartialFunction[(Thread, Throwable), Unit] = {
-            case (t: Thread, e: Throwable) =>
-              Option(t.getThreadGroup)
-                .getOrElse(Thread.getDefaultUncaughtExceptionHandler)
-                .uncaughtException(t, e)
-          }
-        })
-        thread
-      }
-    }
+  ): ThreadFactory = {
+    val p = org.http4s.internal.threads.ThreadPriority(priority.toInt)
+    org.http4s.internal.threads.threadFactory(
+      name,
+      daemon,
+      p,
+      uncaughtExceptionHandler,
+      backingThreadFactory)
+  }
 
   @deprecated("Use newDaemonPool instead", "0.15.7")
   private[http4s] def newDefaultFixedThreadPool(
@@ -66,39 +52,16 @@ object threads {
       name: String,
       min: Int = 4,
       cpuFactor: Double = 3.0,
-      timeout: Boolean = false): ThreadPoolExecutor = {
-    val cpus = Runtime.getRuntime.availableProcessors
-    val exec = new ThreadPoolExecutor(
-      math.max(min, cpus),
-      math.max(min, (cpus * cpuFactor).ceil.toInt),
-      10L,
-      TimeUnit.SECONDS,
-      new LinkedBlockingQueue[Runnable],
-      threadFactory(i => s"$name-$i", daemon = true)
-    )
-    exec.allowCoreThreadTimeOut(timeout)
-    exec
-  }
+      timeout: Boolean = false): ThreadPoolExecutor =
+    org.http4s.internal.threads.newDaemonPool(name, min, cpuFactor, timeout)
 
   private[http4s] def newDaemonPoolExecutionContext(
       name: String,
       min: Int = 4,
       cpuFactor: Double = 3.0,
       timeout: Boolean = false): ExecutionContext =
-    ExecutionContext.fromExecutorService(newDaemonPool(name, min, cpuFactor, timeout))
+    org.http4s.internal.threads.newDaemonPoolExecutionContext(name, min, cpuFactor, timeout)
 
-  private[http4s] def newBlockingPool(name: String): ExecutorService = {
-    val CorePoolSize = 0
-    val MaxPoolSize = Int.MaxValue
-    val KeepAliveTime = 1.minute
-
-    new ThreadPoolExecutor(
-      CorePoolSize,
-      MaxPoolSize,
-      KeepAliveTime.toSeconds,
-      TimeUnit.SECONDS,
-      new SynchronousQueue[Runnable](false),
-      threadFactory(i => s"$name-$i", daemon = true)
-    )
-  }
+  private[http4s] def newBlockingPool(name: String): ExecutorService =
+    org.http4s.internal.threads.newBlockingPool(name)
 }
