+++ testing/src/test/scala/org/http4s/Http4sSpec.scala
@@ -15,8 +15,8 @@ import cats.implicits._
 import fs2._
 import fs2.text._
 import java.util.concurrent.{ScheduledExecutorService, ScheduledThreadPoolExecutor, TimeUnit}
+import org.http4s.internal.threads.{newBlockingPool, newDaemonPool, threadFactory}
 import org.http4s.laws.discipline.ArbitraryInstances
-import org.http4s.util.threads.{newBlockingPool, newDaemonPool, threadFactory}
 import org.scalacheck._
 import org.scalacheck.util.{FreqMap, Pretty}
 import org.specs2.ScalaCheck
